Change type to Float

+++ tb-sample-report-pipeline-plugin/src/main/java/ca/corefacility/bioinformatics/irida/plugins/SnippyPluginUpdater.java
@@ -292,10 +292,10 @@ public class SnippyPluginUpdater implements AnalysisSampleUpdater {
 
 				if (line.contains("in total")) {
 					String[] lineParts = line.split(" ");
-					totalGoodSequences = Integer.parseFloat(lineParts[0]);
+					totalGoodSequences = Float.parseFloat(lineParts[0]);
 				} else if (line.contains("mapped")) {
 					String[] lineParts = line.split(" ");
-					mappedSequences = Integer.parseFloat(lineParts[0]);
+					mappedSequences = Float.parseFloat(lineParts[0]);
 				}
 			}
